requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.1.0
protobuf==4.21.12
Werkzeug==3.1.6
